refactor(application): split managed prose region tests from the compiler suite
The compiler suite covered two responsibilities at once. Managed heading region
and host-prose policy behaviour now lives in its own file, keeping both within
the complexity budget.
